requirements.txt
discord.py==1.2.4
pyyaml==5.1.2
toml==0.10.0
